Add #include directive for megaAVR dependency
The TCB_t type used in the megaAVR version of the library is declared via avr/io.h, but the Servo library does not contain an #include directive for this file. This is not a problem when the library is used in a .ino file, since Arduino.h does have an #include directive for that file and the sketch preprocessor should insert the #include directive for Arduino.h. However, if the Servo library is instead used from a separate translation unit, the dependency was missing and compilation for megaAVR fails:

E:\arduino\libraries\Servo\src/megaavr/ServoTimers.h:36:17: error: 'TCB_t' does not name a type

 static volatile TCB_t* _timer =

                 ^