UNOR4 - update timer period
Instead of fixed 100us timer, it now changes the timer period on the fly
to get better granularity.

And like most other Servo implementations, It also only starts one servo at a time, to help minimize current surge to try to start all of the servos at the same time.

I have done some testing with one and two servos with the sweep, where the 2nd servo is 180-pos...